+++ packages/renderer/src/nodes/applicationCard.tsx
@@ -55,7 +55,11 @@ export interface Icon {
 
 export interface AppCardAction {
   title: string;
-  target: string;
+  target: {
+    app?: string;
+    key: string;
+  };
+  parameters?: object;
 }
 
 export interface AppCardViewProps {